Undo circular dependency
